# Conflicts:
#	src/Riok.Mapperly/Configuration/AttributeDataAccessor.cs
#	src/Riok.Mapperly/Configuration/MapperConfigurationReader.cs
#	src/Riok.Mapperly/Descriptors/DescriptorBuilder.cs
#	src/Riok.Mapperly/Descriptors/ExternalMappings/ExternalMappingsExtractor.cs
#	src/Riok.Mapperly/Helpers/IncrementalValuesProviderExtensions.cs
#	src/Riok.Mapperly/MapperGenerator.cs

+++ test/Riok.Mapperly.Tests/_snapshots/UnsafeAccessorTest.PrivatePropertyInGenericClassMultiple#Mapper.g.verified.cs
@@ -15,7 +15,7 @@ public partial class Mapper
     partial global::B<float> Map(global::A<float> source)
     {
         var target = new global::B<float>();
-        BAccessor<int>.SetValue(target, AAccessor<int>.GetValue(source));
+        BAccessor<float>.SetValue(target, AAccessor<float>.GetValue(source));
         return target;
     }
 }
